+++ toys-core/lib/toys/cli.rb
@@ -29,8 +29,6 @@
 
 require "logger"
 
-require "toys/utils/terminal"
-
 module Toys
   ##
   # A Toys-based CLI.
@@ -65,7 +63,17 @@ module Toys
     #     may use such a file to define auxiliary Ruby modules and classes that
     #     used by the tools defined in that directory.
     # @param [Array] middleware_stack An array of middleware that will be used
-    #     by default for all tools loaded by this CLI.
+    #     by default for all tools loaded by this CLI. If not provided, uses
+    #     {Toys::CLI.default_middleware_stack}.
+    # @param [Toys::Utils::ModuleLookup] helper_lookup A lookup for well-known
+    #     helper modules. If not provided, uses
+    #     {Toys::CLI.default_helper_lookup}.
+    # @param [Toys::Utils::ModuleLookup] middleware_lookup A lookup for
+    #     well-known middleware classes. If not provided, uses
+    #     {Toys::CLI.default_middleware_lookup}.
+    # @param [Toys::Utils::ModuleLookup] template_lookup A lookup for
+    #     well-known template classes. If not provided, uses
+    #     {Toys::CLI.default_template_lookup}.
     # @param [Logger,nil] logger The logger to use. If not provided, a default
     #     logger that writes to `STDERR` is used.
     # @param [Integer,nil] base_level The logger level that should correspond
@@ -77,15 +85,11 @@ module Toys
     #     Default is a {Toys::CLI::DefaultErrorHandler} writing to the logger.
     #
     def initialize(
-      binary_name: nil,
-      config_dir_name: nil,
-      config_file_name: nil,
-      index_file_name: nil,
-      preload_file_name: nil,
-      middleware_stack: nil,
-      logger: nil,
-      base_level: nil,
-      error_handler: nil
+      binary_name: nil, middleware_stack: nil,
+      config_dir_name: nil, config_file_name: nil,
+      index_file_name: nil, preload_file_name: nil,
+      helper_lookup: nil, middleware_lookup: nil, template_lookup: nil,
+      logger: nil, base_level: nil, error_handler: nil
     )
       @logger = logger || self.class.default_logger
       @base_level = base_level || @logger.level
@@ -95,10 +99,13 @@ module Toys
       @config_file_name = config_file_name
       @index_file_name = index_file_name
       @preload_file_name = preload_file_name
+      @helper_lookup = helper_lookup || self.class.default_helper_lookup
+      @middleware_lookup = middleware_lookup || self.class.default_middleware_lookup
+      @template_lookup = template_lookup || self.class.default_template_lookup
       @loader = Loader.new(
-        index_file_name: index_file_name,
-        preload_file_name: preload_file_name,
-        middleware_stack: @middleware_stack
+        index_file_name: index_file_name, preload_file_name: preload_file_name,
+        helper_lookup: @helper_lookup, template_lookup: @template_lookup,
+        middleware_lookup: @middleware_lookup, middleware_stack: @middleware_stack
       )
       @error_handler = error_handler || DefaultErrorHandler.new
     end
@@ -236,6 +243,9 @@ module Toys
                     index_file_name: @index_file_name,
                     preload_file_name: @preload_file_name,
                     middleware_stack: @middleware_stack,
+                    helper_lookup: @helper_lookup,
+                    middleware_lookup: @middleware_lookup,
+                    template_lookup: @template_lookup,
                     logger: @logger,
                     base_level: @base_level,
                     error_handler: @error_handler)
@@ -320,6 +330,35 @@ module Toys
           [:show_help, fallback_execution: true],
           [:add_verbosity_flags]
         ]
+      end
+
+      ##
+      # Returns a default ModuleLookup for helpers that points at the
+      # StandardHelpers module.
+      #
+      # @return [Toys::Utils::ModuleLookup]
+      #
+      def default_helper_lookup
+        Utils::ModuleLookup.new.add_path("toys/standard_helpers")
+      end
+
+      ##
+      # Returns a default ModuleLookup for middleware that points at the
+      # StandardMiddleware module.
+      #
+      # @return [Toys::Utils::ModuleLookup]
+      #
+      def default_middleware_lookup
+        Utils::ModuleLookup.new.add_path("toys/standard_middleware")
+      end
+
+      ##
+      # Returns a default empty ModuleLookup for templates.
+      #
+      # @return [Toys::Utils::ModuleLookup]
+      #
+      def default_template_lookup
+        Utils::ModuleLookup.new
       end
 
       ##