+++ toys-core/lib/toys/definition/arg.rb
@@ -29,8 +29,6 @@
 
 require "optparse"
 
-require "toys/utils/wrappable_string"
-
 module Toys
   module Definition
     ##