+++ toys-core/lib/toys/utils/wrappable_string.rb
@@ -27,8 +27,6 @@
 # POSSIBILITY OF SUCH DAMAGE.
 ;
 
-require "toys/utils/terminal"
-
 module Toys
   module Utils
     ##