+++ toys/CHANGELOG.md
@@ -1,5 +1,7 @@
 # Release History
 
+### 0.3.8 / TBD
+
 ### 0.3.7.1 / 2018-05-30
 
 * FIXED: Fix crash in system update.